+++ package.json
@@ -60,6 +60,7 @@
     "vite": "^3.1.7",
     "vitest": "^0.24.1",
     "vue": "^3.2.40",
+    "vue-router": "^4.5.1",
     "vue-tsc": "^1.0.6"
   }
 }
