+++ src/index.ts
@@ -5,6 +5,7 @@ import 'bootstrap';
 import { storageAvailable } from './utils/validate-storage';
 import { createApp, h } from 'vue';
 import App from './app.vue';
+import router from './router';
 
 import { analyticsInit } from './services/analytics';
 analyticsInit();
@@ -15,7 +16,6 @@ if (!storageAvailable('localStorage')) {
   alert("Please Update your browser for a better experience, I haven't tested with an outdated browser, so you may have no experience at all xD\n\nHey, I'm doing you a favor");
 }
 
-// eslint-disable-next-line no-new
 const app = createApp({
   mounted() {
     // @ts-expect-error
@@ -25,5 +25,8 @@ const app = createApp({
     return h(App, { ref: 'app' });
   },
 });
+
+// Use the router
+app.use(router);
 
 app.mount('#app');
